raw conversion of FakingInternals specs
